+++ src/App.js
@@ -1,21 +1,33 @@
 import React, { useEffect, useReducer, useMemo, Suspense } from 'react';
 import { BrowserRouter as Router, Route, Switch, Redirect } from 'react-router-dom';
 import Cookies from 'universal-cookie';
-import { appInitialState, appReducer, appContext } from 'components/hooks/app';
-import CombinedContext from 'components/CombinedContext.jsx';
-import AdminRoute from 'components/route/AdminRoute';
-import PrivateRoute from 'components/route/PrivateRoute';
+import { app, func } from 'components/hooks';
+import CombinedContext from 'components/combinedContext';
+import AdminRoute from 'components/route/adminRoute';
+import PrivateRoute from 'components/route/privateRoute';
+import Notification, { useNoti } from 'components/notification';
 import { constant } from 'utils/constant';
 import { routes } from 'routes';
 
 const cookies = new Cookies();
 
 export default function App() {
-	const [appState, dispatchApp] = useReducer(appReducer, appInitialState);
+	const [appState, dispatchApp] = useReducer(app.reducer, app.initialState);
+	const [funcState, dispatchFunc] = useReducer(func.reducer, func.initialState);
+	const { show, ...rest } = useNoti();
 
 	const appContextValue = useMemo(() => {
 		return { appState, dispatchApp };
 	}, [appState, dispatchApp]);
+
+	const funcContextValue = useMemo(() => {
+		return { funcState, dispatchApp };
+	}, [funcState, dispatchApp]);
+
+	// init function
+	useEffect(() => {
+		dispatchFunc({ type: constant.SET_NOTIFICATION, notification: show });
+	}, []);
 
 	useEffect(() => {
 		validateCookie();
@@ -46,7 +58,12 @@ export default function App() {
 
 	return (
 		// send array of context to custom component to create layers of context
-		<CombinedContext contexts={[{ context: appContext, value: appContextValue }]}>
+		<CombinedContext
+			contexts={[
+				{ context: app.context, value: appContextValue },
+				{ context: func.context, value: funcContextValue },
+			]}
+		>
 			<Suspense fallback={<div className="loader" />}>
 				<Router>
 					<Switch>
@@ -67,6 +84,7 @@ export default function App() {
 						{/* Admin Route with TopNav and Sidebar*/}
 						<Route path="/admin">
 							<AdminRoute logout={logout} validateCookie={validateCookie} />
+							<Notification {...rest} />
 						</Route>
 						{/* capture invalid route */}
 						<Route render={() => <Redirect to={{ pathname: '/404' }} />} />