+++ src/components/hooks/index.js
@@ -1,0 +1,10 @@
+import AppHook from 'components/hooks/app';
+import FunctionHook from 'components/hooks/function';
+
+/**
+ * Grouping all hook with initial state, reducer, context
+ * and export at once
+ */
+
+export const app = AppHook;
+export const func = FunctionHook;