+++ src/utils/constant.js
@@ -3,4 +3,5 @@ export const constant = {
 	APP: '_o9jqhIjZu6B0kju1.wdAMlj9sO2UYGyKS',
 	SET_LOADING: 'SET_LOADING',
 	SET_LOGIN: 'SET_LOGIN',
+	SET_NOTIFICATION: 'SET_NOTIFICATION',
 };