+++ src/views/User/UserTable.jsx
@@ -1,19 +1,30 @@
 /* eslint-disable react/prop-types */
-import React, { useState, useMemo, useEffect, useContext } from 'react';
-import { appContext } from 'components/hooks/app';
-import EditUserModal from 'views/User/EditUserModal';
-import CustomReactTable from 'components/table/CustomReactTable';
-import DeleteModal from 'components/modal/DeleteModal';
-import Notification, { useNoti } from 'components/notification/Notification';
+import React, { useState, useEffect, useContext } from 'react';
+import { app, func } from 'components/hooks';
+import EditUserModal from 'views/user/editUserModal';
+import CustomReactTable, { useReactTable } from 'components/react-table';
+import DeleteModal, { useDeleteModal } from 'components/modal/deleteModal';
 import { constant } from 'utils/constant';
 
 export default function UserTable() {
-	const { dispatchApp } = useContext(appContext);
-	const { Noti, ...rest } = useNoti();
-	const [showDeleteModal, setShowDeleteModal] = useState(false);
+	const { dispatchApp } = useContext(app.context);
+	const { funcState } = useContext(func.context);
+	const deleteModal = useDeleteModal();
+	const reactTable = useReactTable();
 	const [showEditModal, setShowEditModal] = useState(false);
 	const [user, setUser] = useState({});
 	const [error, setError] = useState('');
+
+	useEffect(() => {
+		deleteModal.init({
+			title: 'Delete User',
+			description: `Confirm to delete user ${user.name}?`,
+			deleteFunction: () => {
+				funcState.notification(`Delete user ${user.name} success`);
+				deleteModal.setShow(false);
+			},
+		});
+	}, [user]);
 
 	// eslint-disable-next-line no-unused-vars
 	const [data, setData] = useState([
@@ -23,53 +34,46 @@ export default function UserTable() {
 		{ id: 4, name: 'David', age: 33 },
 	]);
 
-	const columns = useMemo(
-		() => [
-			{
-				Header: 'Actions',
-				disableGlobalFilter: true,
-				disableSortBy: true,
-				style: { width: 1 },
-				Cell: ({ row }) => {
-					return (
-						<div className="d-flex">
-							<button
-								className="btn btn-warning"
-								onClick={() => {
-									setError('');
-									setUser(row.original);
-									setShowEditModal(true);
-								}}
-							>
-								Edit
-							</button>
-							<button
-								className="btn btn-danger"
-								onClick={() => {
-									setUser(row.original);
-									setShowDeleteModal(true);
-								}}
-							>
-								Delete
-							</button>
-						</div>
-					);
-				},
+	const columns = [
+		{
+			header: 'Actions',
+			accessor: 'actions',
+			disableGlobalFilter: true,
+			disableSortBy: true,
+			style: { width: 1 },
+			Cell: ({ row }) => {
+				return (
+					<div className="d-flex">
+						<button
+							className="btn btn-warning"
+							onClick={() => {
+								setError('');
+								setUser(row.original);
+								setShowEditModal(true);
+							}}
+						>
+							Edit
+						</button>
+						<button
+							className="btn btn-danger"
+							onClick={() => {
+								setUser(row.original);
+								deleteModal.setShow(true);
+							}}
+						>
+							Delete
+						</button>
+					</div>
+				);
 			},
-			{ Header: 'ID', accessor: 'id', id: 'id' },
-			{ Header: 'Name', accessor: 'name' },
-			{
-				Header: 'Age',
-				accessor: 'age',
-			},
-		],
-		[]
-	);
-
-	const deleteUser = () => {
-		Noti(`Delete user ${user.name} success`);
-		setShowDeleteModal(false);
-	};
+		},
+		{ header: 'ID', accessor: 'id' },
+		{ header: 'Name', accessor: 'name' },
+		{
+			header: 'Age',
+			accessor: 'age',
+		},
+	];
 
 	const editUser = (data, reset) => {
 		if (Math.random() < 0.5) {
@@ -80,7 +84,7 @@ export default function UserTable() {
 			setError('');
 			setShowEditModal(false);
 			setUser({});
-			Noti(`Edit user ${user.name} success`);
+			funcState.notification(`Edit user ${user.name} success`);
 		} else {
 			setError('Randomize true false');
 			return false;
@@ -96,7 +100,8 @@ export default function UserTable() {
 		async function getUserData() {
 			setTimeout(() => {
 				if (mounted) {
-					// setData([...data, { id: 5, name: 'New Data', age: 33 }])
+					setData([...data, { id: 5, name: 'New Data', age: 33 }]);
+					reactTable.setTotalRow(data.length + 1);
 				}
 			}, 2000);
 		}
@@ -106,11 +111,26 @@ export default function UserTable() {
 		return () => {
 			mounted = false;
 		};
+	}, [reactTable.pageSize, reactTable.searchInput, reactTable.sortBy]);
+
+	useEffect(() => {
+		reactTable.init(columns, data, {
+			...reactTable.options,
+			title: 'User Table',
+			exportBtn: () => {
+				console.log('export');
+			},
+			filterBtn: () => {
+				console.log('filter');
+			},
+			resetBtn: () => {
+				console.log('reset');
+			},
+		});
 	}, [data]);
 
 	return (
 		<>
-			<Notification {...rest} />
 			<EditUserModal
 				show={showEditModal}
 				setShow={setShowEditModal}
@@ -118,18 +138,8 @@ export default function UserTable() {
 				editUser={editUser}
 				error={error}
 			/>
-			<DeleteModal
-				show={showDeleteModal}
-				setShow={setShowDeleteModal}
-				deleteFunction={deleteUser}
-				title={'Delete User'}
-				description={`Confirm to delete user ${user.name}?`}
-			/>
-			<CustomReactTable
-				data={useMemo(() => data, [data])}
-				columns={columns}
-				// hiddenColumns={['id']}
-			/>
+			<DeleteModal {...deleteModal} />
+			<CustomReactTable {...reactTable} />
 		</>
 	);
 }