+++ src/views/User/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import UserTable from 'views/User/UserTable';
+import UserTable from 'views/user/userTable';
 
 export default function User() {
 	return (